+++ src/api/kakaoApi.ts
@@ -1,8 +1,6 @@
-import { API_PATH, CLIENT_PATH } from '@/constants/path.ts';
+import { API_PATH } from '@/constants/path.ts';
 
 const kakaoApiKey = import.meta.env.VITE_KAKAO_API;
-const localUrl = import.meta.env.VITE_API_LOCAL_URL;
-const kakaoClient = import.meta.env.VITE_KAKAO_CLIENT_ID;
 
 export const getAddress = async (x: number, y: number) => {
   try {
@@ -46,21 +44,5 @@ export const getSearchList = async (query: string, x: string, y: string) => {
     return result;
   } catch (err) {
     console.error(err);
-  }
-};
-
-export const getKakaoInga = async () => {
-  const isAndroid = /Android/.test(navigator.userAgent);
-  // const isiOS = /(iPhone|iPad|iPod)/.test(navigator.userAgent);
-  try {
-    if (isAndroid) {
-      window.location.href = `https://kauth.kakao.com/oauth/authorize?client_id=${kakaoClient}&redirect_uri=${localUrl + CLIENT_PATH.LOGIN_LOADING}&response_type=code`;
-    } else {
-      await window.Kakao.Auth.authorize({
-        redirectUri: localUrl + CLIENT_PATH.LOGIN_LOADING,
-      });
-    }
-  } catch (error) {
-    console.log(error);
   }
 };
