+++ src/domains/Login/Page/index.tsx
@@ -1,13 +1,14 @@
 import { useSelector } from 'react-redux';
 import { RootState } from '@/store';
-import styled from 'styled-components';
 import { useNavigate } from 'react-router-dom';
-import { getKakaoInga } from '@/api/kakaoApi.ts';
 import { BackButton } from '@/components/common/Layout/Header/Header.style.ts';
 import TaxiIcon from '@/assets/icons/login/taxi-icon.svg?react';
 import ArrowLeftIcon from '@/assets/icons/common/arrow-left-icon.svg?react';
 
-const LoginPage = () => {
+import { Header, IconWrap, LoginButton } from './Login.style.ts';
+import { getKakaoInga } from '../api/getKakaoInga.ts';
+
+const Index = () => {
   const navigate = useNavigate();
   const isLogin = useSelector((state: RootState) => state.userSlice.isLogin);
 
@@ -38,35 +39,4 @@ const LoginPage = () => {
   );
 };
 
-export default LoginPage;
-
-const IconWrap = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  font-size: var(--font-large);
-
-  span {
-    font-weight: var(--weight-bold);
-  }
-`;
-
-const LoginButton = styled.button`
-  width: 100%;
-  position: fixed;
-  bottom: 60px;
-
-  & > img {
-    margin: 0 auto;
-  }
-`;
-
-const Header = styled.header`
-  padding: 10px 20px;
-  position: absolute;
-  top: 0;
-  height: var(--header-height);
-  display: flex;
-  align-items: center;
-`;
+export default Index;
